create separate //tests/behaviour/util package
