+++ tests/behaviour/concept/thing/attribute/BUILD
@@ -47,7 +47,7 @@ typedb_behaviour_py_test(
     deps = [
         "//:client_python",
         "//tests/behaviour:context",
-        "//tests/behaviour:util",
+        "//tests/behaviour/util:util",
         "//tests/behaviour/config:parameters",
         "//tests/behaviour/background",
     ],
