+++ tests/behaviour/connection/database/database_steps.py
@@ -27,7 +27,7 @@ from hamcrest import *
 from typedb.client import *
 from tests.behaviour.config.parameters import parse_list
 from tests.behaviour.context import Context
-from tests.behaviour.util import assert_collections_equal
+from tests.behaviour.util.util import assert_collections_equal
 
 
 def create_databases(context: Context, names: List[str]):
