+++ tests/behaviour/connection/transaction/BUILD
@@ -40,12 +40,12 @@ typedb_behaviour_py_test(
         "//tests/behaviour/connection/database:steps",
         "//tests/behaviour/connection/session:steps",
         "//tests/behaviour/typeql:steps",
+        "//tests/behaviour/util:util_steps",
     ],
     deps = [
         "//:client_python",
-        "//tests/behaviour/config:parameters",
         "//tests/behaviour:context",
-        "//tests/behaviour:util",
+        "//tests/behaviour/config:parameters",
         "//tests/behaviour/background",
     ],
     native_typedb_artifact = "//tests:native-typedb-artifact",
