+++ tests/behaviour/typeql/language/undefine/BUILD
@@ -38,7 +38,7 @@ typedb_behaviour_py_test(
     deps = [
         "//:client_python",
         "//tests/behaviour:context",
-        "//tests/behaviour:util",
+        "//tests/behaviour/util:util",
         "//tests/behaviour/config:parameters",
         "//tests/behaviour/background",
     ],
