+++ tests/behaviour/util/util_steps.py
@@ -19,17 +19,11 @@
 # under the License.
 #
 
+from time import sleep
 from behave import *
-
 from tests.behaviour.context import Context
 
 
-@step("connection has been opened")
-def step_impl(context: Context):
-    assert context.client and context.client.is_open()
-
-
-@step("connection does not have any database")
-def step_impl(context: Context):
-    assert len(context.client.databases().all()) == 0
-
+@step("wait {seconds} seconds")
+def step_impl(context: Context, seconds: str):
+    sleep(float(seconds))
